refactor: move participant-related API functions to a dedicated file
